fuck an overflow we u128

+++ programs/hydra/src/utils/logic/calculation.rs
@@ -57,11 +57,11 @@ pub fn update_inflow_for_mint(
         let shares_diff = (fanout.total_shares as u64)
             .checked_sub(tss)
             .or_arith_error()?;
-        let unstaked_correction = diff
-            .checked_mul(shares_diff)
+        let unstaked_correction = ((diff as u128)
+            .checked_mul(shares_diff as u128)
             .or_arith_error()?
-            .checked_div(tss)
-            .or_arith_error()?;
+            .checked_div(tss as u128)
+            .or_arith_error()? as u64);
         fanout_for_mint.total_inflow += unstaked_correction;
     }
     fanout_for_mint.last_snapshot_amount = current_snapshot;
